+++ app-libs/form-component/src/app-components/PDFPreviewControls/index.ts
@@ -1,0 +1,6 @@
+export { PDFPreviewControls } from './PDFPreviewControls';
+export type {
+  PDFPreviewButtonStyle,
+  PDFPreviewControlsProps,
+  PDFPreviewGenerateResult,
+} from './PDFPreviewControls';
