+++ src/App/frontend/src/components/PDFGeneratorPreview/PDFGeneratorPreview.tsx
@@ -4,6 +4,7 @@ import { PDFPreviewControls } from '@app/form-component';
 
 import { useLaxInstanceId } from 'src/features/instance/InstanceContext';
 import { useCurrentLanguage } from 'src/features/language/LanguageProvider';
+import { useLanguage } from 'src/features/language/useLanguage';
 import { generatePdfPreview } from 'src/utils/pdfPreview/generatePdfPreview';
 
 export function PDFGeneratorPreview({
@@ -15,10 +16,13 @@ export function PDFGeneratorPreview({
 }) {
   const instanceId = useLaxInstanceId();
   const language = useCurrentLanguage();
+  const { langAsString } = useLanguage();
 
   return (
     <PDFPreviewControls
-      title={buttonTitle}
+      title={buttonTitle ? langAsString(buttonTitle) : langAsString('pdfPreview.defaultButtonText')}
+      errorHeading={langAsString('pdfPreview.error')}
+      loadingLabel={langAsString('general.loading')}
       disabled={!instanceId}
       showErrorDetails={showErrorDetails}
       onGenerate={(signal) => generatePdfPreview(instanceId, language, signal)}
